# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task CRUD & Auth API",
        "version": "1.0.0",
        "description": "FlyRank Backend Engineering Track Todo API with Supabase Auth & Bearer JWT Protection"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Supabase JWT access token obtained from /auth/login"
            }
        }
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Root metadata info",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Health diagnostic check",
                "responses": {
                    "200": {
                        "description": "Active"
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "summary": "Register a new user account with Supabase",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                },
                                "required": ["email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "User created successfully" },
                    "400": { "description": "Validation error or registration failed" }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Authenticate user credentials & return JWT access token",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                },
                                "required": ["email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Login successful, returns access_token" },
                    "400": { "description": "Missing email or password" },
                    "401": { "description": "Invalid login credentials" }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "summary": "End user session",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "204": { "description": "Logged out successfully" },
                    "401": { "description": "Access token required or invalid" }
                }
            }
        },
        "/public/info": {
            "get": {
                "summary": "Read public open info (No Auth Required)",
                "responses": {
                    "200": { "description": "Success" }
                }
            }
        },
        "/protected/profile": {
            "get": {
                "summary": "Read private profile data (Protected)",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Profile retrieved successfully" },
                    "401": { "description": "Unauthorized - Missing or invalid token" }
                }
            }
        },
        "/protected/dashboard": {
            "get": {
                "summary": "Read user dashboard (Protected)",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Dashboard retrieved successfully" },
                    "401": { "description": "Unauthorized - Missing or invalid token" }
                }
            }
        },
        "/tasks": {
            "get": {
                "summary": "Fetch all task items",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "summary": "Append a new task item",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": { "type": "string" }
                                },
                                "required": ["title"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Created" },
                    "400": { "description": "Validation failure" }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "summary": "Fetch an individual task item by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer" }
                    }
                ],
                "responses": {
                    "200": { "description": "Success" },
                    "404": { "description": "Task not found" }
                }
            },
            "put": {
                "summary": "Update an existing task item",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": { "type": "string" },
                                    "done": { "type": "boolean" }
                                },
                                "required": ["title", "done"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Success" },
                    "400": { "description": "Invalid parameters" },
                    "404": { "description": "Task not found" }
                }
            },
            "delete": {
                "summary": "Delete a task item",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer" }
                    }
                ],
                "responses": {
                    "204": { "description": "No Content" },
                    "404": { "description": "Task not found" }
                }
            }
        },
        "/triage": {
            "post": {
                "summary": "AI Support & Task Triage Classifier",
                "description": "Evaluates unstructured incoming message or task, queries LLM spec with repair retry and Zod schema validation, and returns structured classification JSON.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "description": "Unstructured customer message or task description (1-2000 chars)",
                                        "example": "I need help with my monthly invoice, I was charged twice for pro tier"
                                    }
                                },
                                "required": ["text"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Structured validated JSON triage result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "category": {
                                            "type": "string",
                                            "enum": ["billing", "bug", "feature", "other"]
                                        },
                                        "urgency": {
                                            "type": "string",
                                            "enum": ["low", "normal", "high"]
                                        },
                                        "confidence": {
                                            "type": "number",
                                            "minimum": 0.0,
                                            "maximum": 1.0
                                        },
                                        "reason": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["category", "urgency", "confidence", "reason"]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error (missing field, wrong type, empty or >2000 chars)"
                    },
                    "422": {
                        "description": "Unprocessable Entity (model output failed validation after repair retry, quarantined)"
                    },
                    "503": {
                        "description": "Service Unavailable (LLM kill switch active or provider outage)"
                    },
                    "504": {
                        "description": "Gateway Timeout (LLM call exceeded 30s timeout)"
                    }
                }
            }
        }
    }
}